requirements.txt
streamlit==1.51.0
requests==2.32.5
haystack-ai==2.22.0
dotenv==0.9.9
fastapi==0.123.0
uvicorn==0.38.0
openai-whisper==20250625